fix: resolve clippy linting warnings
- Remove unused import of Stop in required_stops_test
- Remove unnecessary cast from i32 to i32 in job_reader
- Replace useless vec! with array literal in via_stops_test

All clippy warnings resolved.

+++ vrp-pragmatic/src/format/problem/job_reader.rs
@@ -311,7 +311,7 @@ fn read_via_stops(
                             requested_time,
                         )],
                     );
-                    job.dimens.set_via_order(place_idx as i32);
+                    job.dimens.set_via_order(place_idx);
 
                     (job_id, job)
                 })
